Add environment variables to .env.example; Add commands to entrypoint

+++ compose/local/entrypoint.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
-# Run base entrypoint.sh AND delete /compose folder
-bash /compose/base/entrypoint.sh && rm -fr /compose
+# For first time, Run base entrypoint.sh AND then delete /compose folder
+if [ -d /compose ]; then
+	bash /compose/base/entrypoint.sh && rm -fr /compose
+fi
 
 # entrypoint.sh of environment
 python manage.py runserver 0.0.0.0:$DJANGO_PORT
